pyproject.toml
[project]
name = "agent-content-pipeline"
version = "0.1.0"
description = "Bare-metal agent content pipeline for research, synthesis, validation, and content generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7.0,<3.0.0",
  "PyYAML>=6.0.1,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

[tool.setuptools]
include-package-data = true
